Move Util to opekope2.lilac.util
